Tidy up the Graph component (part #2)
